refactor : main.js 코드 구조 분해
