+++ src/utils/routes.js
@@ -1,0 +1,11 @@
+import LoginPage from "../pages/LoginPage";
+import MainPage from "../pages/MainPage";
+import ProfilePage from "../pages/ProfilePage";
+
+const routes = {
+  "/": MainPage,
+  "/login": LoginPage,
+  "/profile": ProfilePage,
+};
+
+export default routes;
